LayoutTests/security/text-decode-long-strings-expected.txt

CONSOLE MESSAGE: stream decoded string from empty array with length 0x7fffffff
CONSOLE MESSAGE: stream decoded string from empty array with length 0x80000000
